Backport input bind_map to different files
libretro/fbalpha@17f11ae
